kati: Ignore TMPDIR when calculating the kati stamp
TMPDIR has to change sometimes between my builds for various complicated
reasons but I think theset also apply to other people who are doing
distributed builds. The result is that the ninja file is regenerated
needlessly. I cannot think of a reason why it *should* be included
therefore it seems reasonable to remove it from consideration.

Change-Id: Ib761b65bba72b1e443b16dd9b2c6fb0d05c7bf5d
Signed-off-by: mydongistiny <jaysonedson@gmail.com>